// src/interfaces/formatter-options.ts
export interface IFormatterOptions {
  indent_size: string
  indent_char: string
  max_preserve_newlines: number
  preserve_newlines: boolean
  keep_array_indentation: boolean
  break_chained_methods: boolean
  indent_scripts: string
  brace_style: string
  space_before_conditional: boolean
  unescape_strings: boolean
  jslint_happy: boolean
  end_with_newline: boolean
  wrap_line_length: string
  indent_inner_html: boolean
  comma_first: boolean
  e4x: boolean
  indent_empty_lines: boolean
}